codership/mysql-wsrep-bugs#1297 inconsistency due to UK column updates
Earlier fix for this issue (bc6c41d463b753c3cc7808f835744810c66f1c82) has caused a regression,
which happens when MDL locks are released and MDL lock waiters are rescheduled. During the rescheduling,
it can happen that the lock releaser ends up BF aborting itself. This is because the lock releaser still holds the MDL lock,
and it is now rescheduling other MDL locks and executing BF aborting on their behalf.

The fix suggested in this commit, skips BF aborting in rescheduling phase, if the BF aborted victim would be self.