codership/mysql-wsrep-bugs#1362 multi table update on table with triggers
Created a mtr test causing crash if multi tabe update is executed on table used for join (read only) and has triggers.
The test renames the triggered table while the update is replicating, this causes the crash if applier cannot open the triggered table, which was just renamed.

Implemented a "naive fix" for the issue.
In this design, the excessive tablemap event remains in binlog, but applier and replayer are relaxed to ignore failure in opening a table.
This means they ignore all open table errors, regardlesss if the failure in open was related to excessive triggered table opening or not.
The assumption is that, if the table open error is a real problem, then applying will fail later on with some other error.

Fix to make sure that ignored table does not remain in table cache
addition to mtr test to verify that table cache does not contains remnants of the zombie table

Fix to raise error and crash the node, if applying fails to open all tables

fixed test galera_trigger_binlog to be deterministic

added two new test cases

Changed the fix to be conditional and configured by wsrep_mode.
Added new wsrep_mode: APPLIER_IGNORE_MISSING_TABLE, which enables the
fix to skip failures with table open at applying time, but only if
the log event contains also other tables, which are sucessfully opened.
APPLIER_IGNORE_MISSING_TABLE option is set as default mode in this commit.

The fix is now extended to raise error, if there is a real inconsistency in the database,
so that a table is really missing, and a multi table operation (like UPDATE)
using also the missing table.

Recorded new result for galera_defaults test, as wsrep_mode has new default value: APPLIER_IGNORE_MISSING_TABLE