Recover wsrep position even if wsrep is off
Before this commit, starting server with --wsrep-recovery
returned random position from uninitialized memory. This was
because the position was not read from storage engine and
the out parameters were not set in wsrep_get_SE_checkpoint().

In order to fix this, removed the check for WSREP_ON at the
beginning of wsrep_get_SE_checkpoint() to run the recovery
regardless if wsrep_on is ON or OFF.

Also fixed casting error in trx_sys_read_wsrep_checkpoint()
to read the stored formatID=-1 as -1 instead of UINT_MAX.
This was required to avoid logging 'Read non-wsrep XID from
storage engines' warning.