Ensure that all output from wsrep_diag.sql goes to same stream
When running wsrep_diag.sql from scripts, output from echo and
SELECTs went into differet output streams. This made the output
hard to read becase the comments echoed with 'echo' were not
ordered with rest of the output.

Changed echo to SELECT 'msg' AS '' to make output easier to
read.