codership/mysql-wsrep-bugs#1549 Fix read-only toggling off under load
This fix was backported from 8.0-v26.

If the server was started with read-only=ON option, the read-only
mode toggled off during concurrent load. The reason was in
`wsrep_read_only_option()`, which modified global opt_readonly
without mutex protection, causing race condition which toggled
read-only permanently off.

This fix removes `wsrep_read_only_option()` call completely.
In addition it being racy, it could not produce correct result
as it inspected lex tree, which was populated for the query
only after the call to `wsrep_read_only_option()` was made.

Added a check for `deny_updates_if_read_only_option()` after
statement execution to avoid retrying autocommit queries which
were rejected because read-only is ON.

Added a MTR test which runs concurrent load on a node which has
read-only=ON and expect all the operations to fail. When the bug
is present, the test will occasionally fail because write
operations succeed despite read-only setting.