codership/mysql-wsrep-bugs#1549 Remove access to lex tree in retrying
The lex tree is not guaranteed to be in well formed state in
wsrep_mysql_parse(). The tree is parsed inside mysql_parse()
which is called from wsrep_mysql_parse(), but the codepath
is relatively complex and the tree may not represent the
actual state after mysql_parse() returns.

In order to avoid accessing lex tree, remove call to
deny_updates_if_read_only_option() in wsrep_mysql_parse().
The condition to replay based on wsrep_conflict_state
should be enough to decide whether autocommit retry should
happen or not.